Refactored/renamed the 'model' package into 'core'
