--- src/model/__init__.py
@@ -1,1 +1,0 @@
-from model.model_class import Model
