+++ src/tk_mvc/__init__.py
@@ -1,4 +1,4 @@
-from .view import View
-from .window import BaseWindow
-from .controller import BaseController
-from .event import Event
+from tk_mvc.view import View
+from tk_mvc.window import BaseWindow
+from tk_mvc.controller import BaseController
+from tk_mvc.event import Event
